+++ src/engine/src/core/Server.py
@@ -379,7 +379,4 @@ class Server:
             logger.info(f"{lbuf('[SERVER]')} ERROR: Failed to resolve idempotency key from provided constraints. {str(e)}. Defaulted to pipeline id '{default_idempotency_key}'")
             return default_idempotency_key
 
-            
-
-